+++ src/pat/depends/depends.js
@@ -4,12 +4,9 @@ import utils from "../../core/utils";
 import logging from "../../core/logging";
 import Parser from "../../core/parser";
 
-// Lazy loading modules.
-let DependsHandler;
-
 const log = logging.getLogger("depends");
-const parser = new Parser("depends");
 
+export const parser = new Parser("depends");
 parser.addArgument("condition");
 parser.addArgument("action", "show", ["show", "enable", "both"]);
 parser.addArgument("transition", "none", ["none", "css", "fade", "slide"]);
@@ -22,8 +19,7 @@ export default Base.extend({
     jquery_plugin: true,
 
     async init($el, opts) {
-        DependsHandler = await import("../../lib/dependshandler");
-        DependsHandler = DependsHandler.default;
+        const DependsHandler = (await import("../../lib/dependshandler")).default; // prettier-ignore
 
         const dependent = this.$el[0];
         const options = parser.parse(this.$el, opts);