+++ src/pat/masonry/masonry.js
@@ -10,12 +10,11 @@ import Base from "../../core/base";
 import utils from "../../core/utils";
 
 // Lazy loading modules.
-let ImagesLoaded;
 let Masonry;
 
-var log = logging.getLogger("pat.masonry");
-var parser = new Parser("masonry");
+const log = logging.getLogger("pat.masonry");
 
+export const parser = new Parser("masonry");
 // parser.addArgument("stagger", "");
 parser.addArgument("column-width");
 parser.addArgument("container-style", '{ "position": "relative" }');
@@ -44,10 +43,8 @@ export default Base.extend({
     trigger: ".pat-masonry",
 
     async init($el, opts) {
-        Masonry = await import("masonry-layout");
-        Masonry = Masonry.default;
-        ImagesLoaded = await import("imagesloaded");
-        ImagesLoaded = ImagesLoaded.default;
+        Masonry = (await import("masonry-layout")).default;
+        const ImagesLoaded = (await import("imagesloaded")).default;
 
         this.options = parser.parse(this.$el, opts);
         // Initialize